requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.1
colorama==0.4.6
et-xmlfile==1.1.0
execnet==2.1.1
Faker==27.4.0
h11==0.14.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PySocks==1.7.1
pytest==8.3.2
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
selenium==4.23.1
selenium-page-factory==2.7
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
